runtime: add SchedStats API
DO NOT SUBMIT. This is an experimental API.

This introduces a runtime.SchedStats API to mirror the existing
runtime.MemStats API. Currently, SchedStats reports the number of
goroutines in four major states: running, runnable, non-go
(syscall/cgo), and blocked.

The intent is that these can be used to determine the CPU load of a Go
process and use this to perform load shedding. This is *not* a
complete solution since the Go scheduler cannot account for threads in
syscalls or cgo; however, a complete solution can be built by
combining these statistics with kernel-provided statistics. The
comments on SchedStats attempt to make this clear.

ReadSchedStats collects these counts efficiently by scanning the P
states and using a running count of the number of goroutines in
syscalls that don't own a P (which avoids doing any additional
accounting in the syscall fast path). This way, it can avoid scanning
all of the goroutines, which could potentially be expensive. With this
approach, at GOMAXPROCS=4, ReadSchedStats takes only 33 ns.

Updates #15490, #17089.

Change-Id: I202f33eea5d10c83dbf41cb45c8c619ff17fa4c4